Changed name from MyArrayList?
I changed the name from MyArrayList (I think that was the name) because I have been changing everything over to table and the old name didn't work for the format.  I did most of the changes in this area.  Added tons of more methods, ways of loading ProjectGUI, and more.