Gregorian Calendar manipulation
I had this with another project.  It contains ways of manipulating Gregorian calendar dates and I figured we can add something to this to help with reminders.  It is pretty well commented which method does what.  We can get rid of any methods we don't use.  I've already used 2 or them I think.