Now new, edit, sub GUI
I made some serious changes.  The big one?  The project no longer uses the subGUI.  I just couldn't get it to print to projectGUI the way that it was.  CreateGUI now handles adding a new project, editing a project, and adding a sub group.  Check it out.  It changes and adjusts to what is needed.