IsValidField
Added a date checker to prevent new project due date from being before today's date. Thought we already had this and maybe it wasn't working because me changing daysLapsed, but I'm really tired and can't find it so I made a thing. If we don't need it we can get rid of it and change the other code to work or whatever but it is 5:33 am and I'm going to bed. Project looks good. Peace out.